Add collector.ProcessIndividualSpan
When meta.refinery.individual_span is set on a span, make an immediate
decision using the configured samplers as if the span were its own
trace. Do not save the decision for the trace id (so other spans from
the same trace also get handled individually). This enables using the
EMA samplers to preserve the most interesting events rather than the
most interesting traces (which may contain a lot of other unrelated
things we care less about).