Key batch sampling on sampler and key, improve metrics
We want to avoid grouping keys by different sampler rules so that the
last thing with a given key (which is generated by sampler rules)
doesn't clobber the entire batch. This still won't play that nicely with
deterministic samplers that are not either 0% or 100%.

Fix cache size metric to account for number of traces instead of number
of batches (but record those as well).

Add debugging fields to spans emitted via batch sampling.

Export counts from all histograms.